pyproject.toml
[tool.poetry]
name = "larimar"
version = "0.1.0"
description = "1YES!"
authors = ["larimarit <larimaritgroup.ru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.6.0"
astroid = "^2.14.2"
async-timeout = "4.0.2"
attrs = "21.4.0"
Automat = "20.2.0"
autopep8 = "1.5.1"
beautifulsoup4 = "4.8.2"
certifi = "2020.4.5.1"
cffi = "1.15.0"
chardet = "3.0.4"
colorama = "^0.4.6"
constantly = "15.1.0"
cryptography = "36.0.1"
distlib = "0.3.6"
Django = "4.1.7"
django-bootstrap-datepicker-plus = "^5.0.3"
django-bootstrap4 = "1.1.1"
django-ckeditor = "^6.5.1"
django-crispy-forms = "^2.0"
django-group-by = "0.3.1"
django-jquery = "3.1.0"
django-js-asset = "^2.0.0"
django-mptt = "0.14.0"
django-mptt-admin = "2.4.1"
django-tables2 = "2.3.1"
djangorestframework = "3.12.4"
docopt = "0.6.2"
filelock = "3.9.0"
hyperlink = "21.0.0"
idna = "2.9"
importlib-metadata = "4.11.3"
incremental = "21.3.0"
isort = "4.3.21"
lazy-object-proxy = "1.4.3"
lxml = "4.5.1"
Markdown = "3.3.4"
mccabe = "0.6.1"
msgpack = "1.0.3"
msgpack-python = "0.5.6"
pipenv = "2022.12.19"
pipreqs = "0.4.10"
platformdirs = "2.6.2"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycodestyle = "2.5.0"
pycparser = "2.21"
pylint = "^2.16.2"
pylint-django = "^2.5.3"
pylint-plugin-utils = "^0.7"
pyOpenSSL = "21.0.0"
python-dateutil = "2.8.1"
python-dotenv = "0.21.1"
pytz = "2020.1"
requests = "2.23.0"
sentry-sdk = "1.0.0"
service-identity = "21.1.0"
six = "1.15.0"
soupsieve = "1.9.5"
sqlparse = "0.3.1"
toml = "0.10.1"
txaio = "21.2.1"
typing_extensions = "4.0.1"
urllib3 = "1.25.9"
virtualenv = "20.17.1"
virtualenv-clone = "0.5.7"
wrapt = "^1.14.1"
yarg = "0.1.9"
zipp = "3.8.0"
django-rosetta = "^0.9.8"
mysqlclient = "^2.1.1"

[tool.poetry.group.dev.dependencies]
channels-redis = "^4.0.0"
aioredis = "1.3.1"
hiredis = "2.0.0"
redis = "^4.5.1"
autobahn = "^23.1.2"
Twisted = "^22.10.0"
channels = "3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

